+++ generators/app/templates/config/passport.js
@@ -7,7 +7,8 @@ var passport = require('passport'),
     LocalStrategy = require('passport-local').Strategy,
     JwtStrategy = require('passport-jwt').Strategy,
     FacebookTokenStrategy = require('passport-facebook-token').Strategy,
-    TwitterTokenStrategy = require('passport-twitter-token').Strategy;
+    TwitterTokenStrategy = require('passport-twitter-token').Strategy,
+    YahooTokenStrategy = require('passport-yahoo-token').Strategy;
 
 passport.use(new LocalStrategy({
     usernameField: 'username',
@@ -138,6 +139,43 @@ passport.use(new TwitterTokenStrategy({
             });
     } else {
         req.user.twitter = profile._json;
+        req.user.save(next);
+    }
+}));
+
+passport.use(new YahooTokenStrategy({
+    clientID: "<%= answers['application:yahoo-app-id'] %>",
+    clientSecret: "<%= answers['application:yahoo-app-secret'] %>",
+    passReqToCallback: true
+}, function (req, accessToken, refreshToken, profile, next) {
+    if (!req.user) {
+        User
+            .findOrCreate({
+                'yahoo.id': profile.id
+            }, {
+                username: req.param('username') || profile.username || profile.displayName,
+                email: req.param('email') || (profile.emails && profile.emails[0].value),
+                firstName: req.param('firstName') || (profile.displayName && profile.displayName.split(' ')[0]),
+                lastName: req.param('lastName') || (profile.displayName && profile.displayName.split(' ')[1]),
+                photo: req.param('photo') || ['https://graph.facebook.com/', profile.id, '/picture?type=large'].join(''),
+                facebook: profile._json
+            })
+            .exec(function (error, user) {
+                if (error) {
+                    return next(error);
+                }
+
+                if (!user) {
+                    return next(null, false, {
+                        code: "E_AUTH_FAILED",
+                        message: "Yahoo auth failed"
+                    });
+                }
+
+                return next(null, user);
+            });
+    } else {
+        req.user.yahoo = profile._json;
         req.user.save(next);
     }
 }));