+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improvement: Implemented parsing CommonJS modules from AST (Abstract Syntax Tree), so now generator is knows which modules you are really using;
 - Improvement: Frozen structure of generator and project. Any serious changes in project structure;
+- Improvement: Add Yahoo authorization;
 - Improvement: Split generator arguments declaration to separate folder;
 - Improvement: Split generator options declaration to separate folder;
 - Improvement: Split generator questions declaration to separate folder;